Merge to master

+++ src/users/routes.js
@@ -1,4 +1,5 @@
 const Celebrate = require('celebrate');
+const Sequelize = require('sequelize');
 
 const User = require('./model');
 const schema = require('./schema');
@@ -17,7 +18,12 @@ function createUser(req, res) {
     .then((model) => {
       res.json(model.toJSON());
     })
-    .catch(() => {
+    .catch((err) => {
+      if(err instanceof Sequelize.UniqueConstraintError) {
+        return res.status(400).json({
+          error: 'A user with that email aready exists'
+        });
+      }
       res.status(500).json({
         error: 'There was a problem saving the user'
       });
@@ -28,10 +34,16 @@ function deleteUser(req, res) {
   const userId = req.params.id;
   User.findById(userId)
     .then((model) => {
-      return model.destroy({ force: true });
-    })
-    .then(() => {
-      res.json({});
+      if(!model) {
+        return res.status(404).json({
+          error: 'The user does not exist'
+        });
+      }
+
+      return model.destroy({ force: true })
+        .then(() => {
+          res.json({});
+        });
     })
     .catch(() => {
       res.status(500).json({
@@ -59,6 +71,12 @@ function getUserById(req, res) {
   const userId = req.params.id;
   User.findById(userId)
     .then((model) => {
+      if(!model) {
+        return res.status(404).json({
+          error: 'The user does not exist'
+        });
+      }
+
       res.json(model.toJSON());
     })
     .catch(() => {
@@ -71,12 +89,23 @@ function getUserById(req, res) {
 function updateUser(req, res) {
   User.findById(req.params.id)
     .then((model) => {
-      return model.update(req.body);
+      if(!model) {
+        return res.status(404).json({
+          error: 'The user does not exist'
+        });
+      }
+
+      return model.update(req.body)
+        .then((model) => {
+          res.json(model.toJSON());
+        });
     })
-    .then((model) => {
-      res.json(model.toJSON());
-    })
-    .catch(() => {
+    .catch((err) => {
+      if(err instanceof Sequelize.UniqueConstraintError) {
+        return res.status(400).json({
+          error: 'A user with that email aready exists'
+        });
+      }
       res.status(500).json({
         error: 'There was a problem updating the user'
       });
